conversational-ai/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
groq
pydantic
pypdf